Removed the 'diff from default OPTIONS' logic from DBBenchRunner
Changes per file:
bench_runner: shifted BenchmarkRunner abstract class to this file; also
shifted method get_info_log_file_name to this class from DBBenchRunner
as other bench runners may need it too
db_bench_runner: removed the OPTIONS diff logic; assumption is that
db_bench should be able to handle Rocksdb OPTIONS given as command-line
args even when options_file arg is given; added more comments to the
code
ALL FILES: added user-id to the TODO comments

Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
